backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pymilvus>=2.4.0
milvus-lite>=2.4.0
zhipuai>=2.1.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
pypdf>=4.0.0
python-docx==1.1.2
aiofiles==24.1.0
